Fixes #42:Try integrating announcements/notices into app
