+++ app/src/main/java/set/forum/ForumData.java
@@ -6,12 +6,12 @@ import java.util.List;
  * Created by siddhant on 1/17/17.
  */
 
-public class SiteNews {
+public class ForumData {
 
     public List<Discussion> discussions = null;
     public List<Object> warnings = null;
 
-    public SiteNews() {
+    public ForumData() {
 
     }
