+++ app/src/main/res/layout/row_site_news.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context="crux.bphc.cms.fragments.SiteNewsFragment"
+    tools:context="crux.bphc.cms.fragments.ForumFragment"
     android:layout_margin="8dp">
 
     <LinearLayout
